[LLVM] introduce and convert the LLVMStructType (#133)
So I need the llvm.struct<(i32)> type for my project. Turns out: the llvm dialect is not exposed in the python bindings so I cannot go the usual route. For that reason, I came up with a really mad hack IMO. Feel free to comment if you see any way this could be improved.

Co-authored-by: Fehr Mathieu <mathieu.fehr@gmail.com>