Compiler: fix mlir module loading (#221)
This PR fixes loading of MLIR modules. The behavior we want here is that, if we call load_mlir_module the first time (i.e. when _mlir_module is still None), the given argument is written into _mlir_module (given that the required submodules are loaded). If the function is called again, the given mlir_module argument needs to be the same each time.