CI: Generate proper coverage for filecheck tests (#268)
This PR resolves an oversight in the earlier CI PR. In that version, the source directories were specified in a relative way (which is the only way it really makes sense), but since lit decides to run everything in its subdirectory, these relative paths did not exist. Therefore, all the coverages generated by the filecheck tests were basically empty (no source files) and they were not taken into account.

This is a solution to workaround this issue. The other solution I see would be to try and parametrize the config file somehow, such that an absolute source path can be defined...