frontend: Add constants (#337)
Introduces compile-time constants to the frontend:
```python
a: Const[i32] = 2**8
```
Constants can be used everywhere and must have a unique name (reassignment
is not allowed). This way we do not depend on semantics of `AnnAssign` or
`Assign` and error on both:) At compile time, constants are evaluated and
inlined into AST. In the future, type checks for constants will be added.

Tested using both `lit` ad `pytest`.