CI: Reconfigure CI to cache built binaries (#369)
This commit removes the dependence from the MLIR Python bindings and changes the caching approach. Instead of using ccache an rebuilding, the entire build folder is cached (to get mlir-opt, which we need for differential testing).
Additionally, this changes the way lit decides which tests to run from "are the Python Bindings available" to "is mlir-opt in the path".