misc: Remove usage of coverage in xdsl-opt. (#433)
Use coverage CLI to get coverage from filechecks instead. I open this PR as a draft to make sure codecov report is still working properly first.

This removes the need to have coverage installed to just run xDSL.
Bonus: coverage is the only non-pure dependency left, so if we remove it, xDSL can be used as a pure-Python tool. This means that only a Python distribution is needed to run it, not even some precompiled wheels for some dependencies! This is nice for e.g. JupyterLite, or other exotic platforms.


---------

Co-authored-by: Emilien Bauer <bauer.emilien@gmail.com>