docs: A few typo fixes, plus some comments. (#547)
I'd change "Assembly" to "assembly" throughout.

"Encode domain-specific knowledge (i.e. optimisations)" - "i.e." is not appropriate here as domain-specific knowledge and optimisations are not the same thing.

Under "Operands", can an operand be a block argument? I.e., can we use a block argument in an operation?

In the SQL dialect example, the use of lambda comes out of the blue - perhaps I missed it but I didn't see lambdas mentioned earlier in the more general discussion of xDSL.