core: Add traits  (#578)
This PR adds traits to operations.

### Traits

https://mlir.llvm.org/docs/Traits/

Traits are used to provide operation invariants or additional semantic information.
For instance, we can have a `Terminator` trait, an `HasParent` trait, or a `Pure` trait (this one is actually more complex in MLIR, but we can start with an easy version).

Traits may have parameters, and we currently have no constraints on what these parameters can be.

Traits are currently attached to operation definitions using the `traits` field. Operations will
also inherit parents traits.

```python
@irdl_op_definition
class NoResultsOp(Operation):
    name = "test.no_results"

    traits = frozenset([NoResultsTrait()])
```

### Verification

Traits may provide a `verify` method that will do additional checks on an operation.
These checks are done after the local constraint checks (i.e., checking the existence of type of each operand, result, ...),
but before the custom operation verification function.
This is in line with what MLIR is currently doing. 

### Interfaces

This PR does not include Interfaces. I tried merging interfaces with traits to simplify the design, but it was just
making things more complex, notably because interfaces do not have parameters.

### Design questions

#### How to assign traits
Currently, we assign traits to operation definitions with the `traits` field, which is a class field.
We could also try to do this in the decorator, though I'm not sure if we would gain much from it to be honest.

#### Should we inherit parent traits ?
I'm also unsure about inheriting traits from parent classes. The code does not make it obvious, so maybe
we should opt to this design, which I am also unsure of:
```python
@irdl_op_definition
class NoResultsOp(ParentOp):
    name = "test.no_results"

    traits = ParentOp.traits | frozenset([NoResultsTrait()])
```