RFC: implicit builder regions (#634)
This PR proposes an API for a clutter-free IR building experience. By
introducing some global mutable state, we can construct IRs without
explicit `b.insert` calls. Obviously this is super error-prone, so I
added some asserts in the code to make sure that the right thing is
going on, and try to generate useful errors when something goes wrong.
Please take a look at the tests for usage.