core: deprecate Region.from_op_list (#736)
This is the last deprecation in favour of `Region.__init__` one thing to
watch out for is that `Region.get([])` produces a region with a single
empty block, and `Region([])` produces a region with no blocks. In the
now closed variadic init PR, this meant that the new way to spell
`Region.get([])` was `Region(Block())` I wonder if we want to make
`Region()` create an empty single block region? Or maybe that would be
more confusing? If so, maybe this would be the PR to add it.

I look forward to your feedback!