CI: Retrieve pyright version from requirements file (#766)
Okay, this is done now. The pyright workflows now first parse the
requirements file and write the proper pyright version to a file. This
file is then read in as an environment variable, which is later used to
decide the version.

This means we can now just merge the bumps that dependabot suggests.