core: move Block.ops to a doubly-linked list structure (#762)
This more closely mirrors the MLIR structure. We observed a significant
speedup when rewriting, likely due to the fact that linked lists are
much more efficient at mutation towards the start of the list, avoiding
large copies of elements. I tried to separate out all the changes I
could from this PR but it's still rather long. I have more changes on
top of this that migrate all client code away from using indices for
block insertions, but would rather open them up for review once this is
merged.

---------

Co-authored-by: Michel Weber <55622065+webmiche@users.noreply.github.com>