dialects: add register index and name to riscv.Register (#848)
This lets us to do partial register allocation by assigning an index on
the type of the SSA value. I can confirm that this "works" in the
`sasha/workshop-all` branch, where code in Toy is lowered to RISC-V and
runs, including things like functions & function calls, syscalls, etc.
My understanding is that it works slightly differently in the ChocoPy
project, but I haven't looked at that yet, so would appreciate feedback
on how this approach might be better or worse than that, and other
ideas.