dialects: Improve Stencil Fieldtype (#907)
This PR replaces `from_shape` constructor for `stencil.fieldtype` with
`__init__()`. I have also added methods to get the shape and no. of
dimensions to improve usability and enforce consistency with other types
(`MemRefType` and `VectorType`).

End goal for these modifications is to eventually get the type outside
of experimental directory (will do this in a separate commit).