frontend: split into "core" and "default"
Things in "core" are hardcoded part of the frontend translation (i.e., we do not allow frontends to change the meaning of function definitions, if/else, assignment, and a variety of other core python language features).

However, we do allow frontends to define their own semantics for data types and introduce their own operations (written as function calls in the Python source, i.e., foo.op(a,b)).