frontend: Strengthen Python code check (#802)
TL;DR:
We allow only two ways of creating a front-end program:
```python
def foo():
  # no blocks, ops only!

  # explicit return
  return
```
and
```python
def foo():
  @block
  def bb0():
    # ops only!

    # explicit return
    return

  # explicit entry block call
  return bb0()
```
The only exception are compile-time constants which are inlined into
functions/blocks and therefore do not affect code generation.

Previously, the code check was not great: 
1) Code duplication
2) Not too restrictive and wrong wrt blocks (e.g. blocks did not carry
explicit terminators)
3) Similar to 2), functions were not required to have a terminator
4) We were allowing grouping of ops in functions, and without them

Now, we only allow functions in the program. The reason is: analysis is
simpler, if you want operations then you can just put them inside a
dummy function. Also, both functions and blocks need terminators. This
way we can make more assumptions for code generation, e.g. that the last
statement in the function is a `return`, etc. If we ever want to add
more fancy things (like not specifying void return), this can be
implemented by adapting the AST but keeping code generation logic
intact.

This also fixes basic blocks. Now each block has to return either a call
to another block, and if expression with calls to other blocks, or
otherwise it is considered as a terminating block for a function. In
case 0 or >1 terminating blocks exist, we abort. Again, this make code
generation for blocks (not supported yet) way easier and nicer.

Lastly, there are no real changes to code generation pipeline, only the
checks before. This will be fixed in a next patch so that it is easier
to review.

---------

Co-authored-by: Michel Weber <55622065+webmiche@users.noreply.github.com>