core: Integrate parser for more complex pass pipeline specs into xdsl-opt (#964)
This is the final stage of the pass pipeline parser.

This PR hooks up the parser to xdsl-opt. A pass is now a dataclass that
can have fields:

```python
@dataclass
class CustomPass(ModulePass):
    name = "custom-pass"

    number: int | float

    int_list: list[int]

    non_init_thing: int = field(init=False)

    str_thing: str

    optional: str | None
```

Which can then be called from the command line as `custom-pass{number=1
int_list=1,2,3 str_thing=hello}`.

Optional fields are not required to be present.

*A note on edge cases:*

Due to the nature of the syntax, we can't know if `x=3` means `3` or
`[3]`. So if we see an annotation like `int | list[int]` we cannot know
how to convert. This implementation will always resort to non-lists in
cases like this.

---------

Co-authored-by: Michel Weber <55622065+webmiche@users.noreply.github.com>