dialects: riscv: Add basic register allocator strategy (#995)
The current register allocator for the RISC-V dialect exclusively
utilizes j-registers for unallocated values. However, @compor and I
propose an improvement over the existing allocator by introducing a
naive strategy that leverages the available real RISC-V registers for
each block. In case the real registers are exhausted, the allocator
gracefully falls back to j-registers.

This enhancement brings forth several important questions that need to
be addressed at this stage:

- Determining the Set of Available Registers: The choice of available
registers depends on the ABI and the adopted calling convention. So, we
probably want to determine a target for now and see how it involves
other topics (like function handling see
https://github.com/xdslproject/xdsl/pull/929).

- Efficiency Requirements for Future Work: We might need to establish
the desired level of efficiency for the register allocator (linear scan,
...) and the relevant use cases.

Moreover, this improvement is contingent on the merging of the LabelOp
into the main branch. For further details regarding the LabelOp merge,
please refer to the following pull request
https://github.com/xdslproject/xdsl/pull/994.

---------

Co-authored-by: Chris Vasiladiotis <cvassiladiotis@gmail.com>