core: use with blocks for implicit building (#1017)
One thing that I'm unhappy about with the current implicit builder APIs
is that one has to first define the function that builds the region, and
then construct the operation with that region. This is less than ideal,
as they are pretty closely tied together, so the visual separation can
make things harder to parse. I tried solving this first with
higher-order functions but that got messy pretty quickly.

This PR introduces an alternative to this, which I now think is
superior, and would like to transition to for the whole codebase if we
decide to accept this PR:

Before:
``` python
@Builder.implicit_region(multiply_transpose_type.inputs)
def multiply_transpose(args: tuple[BlockArgument, ...]) -> None:
    a, b = args

...

toy.FuncOp(
    "multiply_transpose",
    multiply_transpose_type,
    multiply_transpose,
    private=True,
)

```

After:
``` python
with toy.FuncOp.implicit_builder(
    "multiply_transpose", multiply_transpose_type, private=True
) as (a, b):
```

The new approach gets rid of Python functions altogether, exposes args
as part of the `with` block API, brings the order in which the
operations are created in Python in line with their order in the IR. If
we migrate all our functions this way, it'll let us have an API where
the user doesn't need to import the Builder at all, only using helpers
on Operations.