dialects: Add element_count to ShapedType and add ContainerType (#1065)
This adds two things:

- an `element_count` property to the `ShapeType`, that is automatically
derived and uses the `get_shape` method internally
- A new `ContainerType` class that allows us to have a common interface
for types that encompass another type (e.g. MemRef, Vector, Tensor,
Dense, llvm pointer)

In the process, I moved most of them over to this. That changed their
shape return value to `tuple` (previously array for some builtins). I
fixed some tests and had to touch the printer a bit sadly.

(This was done in response to me finding a bug in how memref element
counts are calculated in the mpi lowering, which was pretty sad)