dialects: (riscv) add verifiers for riscv ops using immediate (#1027)
In the RISC-V instruction format, the range of immediate values is
constrained, such as a 12-bit immediate for most I-Type instructions.
However, the current implementation of the dialect lacks the necessary
validation for these immediate values. This pull request aims to address
this issue by introducing verifiers for the operations to check that the
values we pass are actually in those ranges.

The main goals of this pull request are as follows:

- We ensure that the immediate values used in RISC-V instructions fall
within the valid range. This verification is quite important if we want
to do some code generation optimizations, such as instruction selection
(e.g. merging load constant and add into addi).

An question arising from this pull request pertains to the handling of
immediates in different RISC-V targets, specifically RV32I and RV64I.
The current implementation assumes a RV32I target for simplicity (but in
the RV64I target some same instructions use larger immediates). So any
feedback on that is appreciated.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>