core: Implement parsing for Affine IR constructs (#1147)
The Parsing is complete even for semi-affine maps. Although, the
AffineMap/AffineExpr IR constructs cannot handle them. The Affine IR
constructs currently cannot handle some things like `(4 * 1) * d0`,
because they cannot understand if `4 * 1` is a constant or not, but this
will be improved in a separate PR with more tests.