misc: reorder test execution order in Makefile (#1200)
This lowers the expected time to see an error, since notebook tests are
the slowest to run, and the least likely to break. I also split out the
toy tests to their own target since that's something I run a lot when
working on that bit of code, all that does is change the order of
execution.