dialects: (riscv) add initial support for F extension  (#1188)
This PR introduces some changes needed to support the handling of
floating point operations (F extension) in the `riscv` dialect and all
the operations themselves, which is necessary for future works.

The main points are:
- Split the class `RegisterType` into two different classes to
distinguish between
integer and floating point registers. This is handy when register
allocating and
also to ensure the soundness of the dialect in order to stay in line
with the ISA.
Thus, the user has to specify directly the type of the register when
using it inside
  an operation (`riscv.ireg` or `riscv.freg`).
- I'm still unsure if that's the best way of doing this, so I'm open to
suggestions.
- I've updated already most of the test suite, including the emulation
tests, where I found some issues in the `riscemu` library which resulted
in a PR (https://github.com/AntonLydike/riscemu/pull/27), still pending
to be merged.