reqs: add typing-extension dependency (#1263)
Self is a very useful type to have available, introduced in 3.11. In
Python 3.10 a third-party module is required for the same functionality.
I reckon the extra dependency is worth it for the kind of thing we'd
like to express.

It seems to me like it's necessary to make it a required dependency as
core code should be able to use Self, so users of pip install should
inherit it as well. It's a bit annoying to have a dependency that's only
required for python 3.10 but it seems worth it to me.

---------

Co-authored-by: kingiler <68145845+kingiler@users.noreply.github.com>