core: add MLIROptPass, calling mlir-opt (#1257)
This adds a Python class to pipe code through `mlir-opt`. This has been
quite useful to me while writing the pass pipeline in Python, to
leverage things like cse and lowering through builtin dialects. I'm
planning to take mlir-opt out of the Toy tutorial for the sake of having
an end-to-end compiler with no dependencies, but I think it will be
useful to future projects using xDSL.