dialects: (riscv) remove Register class (#1281)
When I initially wrote the register types, the reasoning in my mind was
that I wanted an optional register name, and `Data` cannot accept an
optional generic parameter. But actually an empty string seems more
appropriate than None anyway. Now that we have floats in riscv, and some
duplication, it feels like a good idea to simplify. This also includes a
rename of RegisterType to RegisterAttr, since we don't exclusively use
them as types.

After some discussions with @kingiler on the changes in #1269, I thought
I'd play around with the code to see how to make the generics work for
both Register/FloatRegister and RegisterType/FloatRegisterType. There
were always some issues with it locally, and quite a lot of complexity
due to the extra step of indirection. This is what led me to the changes
in this PR, which I propose as an alternative to #1269. It's probably a
good idea for reviewers to take a look at both to have a better view of
multiple potential ways forward.