docs: (toy) add lowering to Affine (#1289)
The long-awaited Toy to builtin lowering. I added some code that in MLIR
is in the affine dialect itself only to Toy for the time being. I think
it's better to have this code in and functional enough for Toy for now,
and move it to the main framework later.