core: Add declarative assembly format specification to IRDL (#1313)
This PR adds the first version of the [declarative assembly
format](https://mlir.llvm.org/docs/DefiningDialects/Operations/#declarative-assembly-format).
In its current form, it only supports the `attr-dict` directive, which
specifies where the attribute dictionary should be parsed.

The structure of the PR is:
* `irdl.py` was modified (and moved to `irdl/irdl.py`) to allow
specifying a declarative format with the `assembly_format` field.
* `declarative_assembly_format_parser.py` will parse the format and
create the corresponding data structures, as well as checking that the
format is correct. The parser is written by reusing the MLIR format
lexer and parser. This is slightly hacky, but makes the change much
smaller.
* `declarative_assembly_format.py` contains the data structure to
represent the declarative format, and to parse/print a particular
format.

Following PRs will incrementally add more features. I couldn't really
remove more from this PR without making it impossible to understand.