misc: disable type ignore (#1400)
Type ignore sometimes hides the wrong issue. We sometimes want to just
get pyright to stop complaining about a private funciton call and then,
when the function interface changes or is deleted, we miss the type
error. When ignoring pyright, it's best to always write `pyright:
ignore[specific, issues]` as opposed to a blanket `type: ignore`. I also
found some `type: ignore[specificIssue]` but that's not actually a
syntax that `type: ignore` supports, it just ignores the whole line.