dialects: (riscv) add fmvop (#1498)
This doesn't really use it anywhere yet, just provides the ability to
express copying a value from one register to another. Hopefully we'll
actually manage to optimise out this operation for all the cases we care
about, but it's possible that we won't so we'll need to implement a
lowering to the rountrip through memory before printing.