backend: (rsicv) remove cast to int hack in riscv lowering (#1462)
We currently have a few hacks in our lowering that forcefully convert
all values to integers. SInce we've recently added float support in a
few places, we can remove that hack. Unfortunately it's not one hack but
a jenga tower of hacks that are all weirdly interpdependent. The PR is
larger than I would have liked but I feel that we might be better off
yanking it out in one go.