documentation: (toy) remove mlir-opt dependency for lowering Toy to riscv (#1532)
Removing the dependency lets us have the full Toy to riscv compiler run
in the jupyter notebooks online, and generally feels like a nice
decouplng to me.

This PR removes a bunch of code that I added only as a workaround for
functionalities that were missing at the time but that we have now
thanks to the work everyone's been doing on the riscv side of things. My
ultimate goal is to remove all the riscv-related code in Toy, as we add
more functionality to the riscv compilation pipeline in xDSL itself.