transformations: move riscv canonicalization file to dedicated folder (#1556)
This is a minor framework organisational change, I think it makes sense
to have a submodule in the transformations module to house all the
canonicalisation patterns. I also renamed the filecheck test that checks
for canonicalization behaviour to just `canonicalize.mlir` which seems
like a good convention to follow, the one for dmp is already called
that.