dialects: Add HLS dialect (#1578)
HLS dialect and HLS transformations. The dialect is an abstraction of
the typical HLS constructs, giving support for interfaces,
task-parallellism and so on. The transformations focus on lowering the
stencil dialect to FPGA through a shift buffer based optimisation.
Currently we have support for PSyclone implemented. Support for Devito
will be added soon.