dialects: (riscv) add frep operations to riscv dialect (#1587)
This is the first sketch of a frep dialect. I think it would make most
sense for frep to take some operands and return some also, something to
keep track of the fact that the stream registers are used correctly.
Maybe these could be riscv registers, maybe a new kind of streaming
register, not sure yet. I think we can do this later as we build up the
compilation flow.