core: test and fix IRDLOptions as properties (#1611)
I forgot something while extracting this from #1582 in #1603; This is
fixed and tested here.

Basically, by switching some helpers to use the IRDLOptions types rather
than placeholder instances, I forgot to actually get the option instance
when using this option in the irdl generated builder, causing this to
always fall back to the attribute-based behaviour and causing errors in
the property-expecting case (hence no regression appearing with no
dialect ops using properties yet). Fixed :wrench:

this is also the last core-preparing PR before #1582 :tada: Opening a
Zulip thread in a moment to discuss that.