interpreter: add mechanism to store runtime data in the interpreter (#1628)
This is necessary for some further interpreter changes. The interpreter
function implementations are suposed to be stateless, and the riscv
interpreter functions currently violate that, by caching the data on the
functions object itself. This PR fixes that by introducing a new cache
on the interpreter object, which is stateful already.