dialects: (seq) Add ClockType and ClockDivider (#1619)
`seq` dialect first part, as discussed on zulip.

Opened as WIP but looking for feedback. Pinging @superlopuh
@tobiasgrosser @compor and @JosseVanDelm as having participated in the
discussion on zulip. Not sure who’s signing off on changes down the
line.
There seems to be quite a lot to discuss just on this first
type/operation
- One thing nagging me in particular is that printing the
`ClockDivider`’s power of 2 attribute with `print_attribute` means we
get `seq.clock_div %clk by #int<4>` instead of the input syntax
`seq.clock_div %clk by 4` -- should I print the naked `int` value to get
the right syntax instead? It seems to be parsed back fine.
- I haven’t managed to coerce this `pow2` attribute to be signless
8-bit, no matter how much I’ve been messing around with `IntAttr`,
`IntegerType`, and other similarly named things. It always seems to be
read by `parse_attribute()` as a 32 or 64 bit int -- suggestions welcome
here.

Some questions that will need answering before any of this can go
forward:

1. What’s the source of truth for a this dialect: CIRCT docs or code?
They disagree in major ways, most notably docs define
[`ClockType`](https://circt.llvm.org/docs/Dialects/Seq/#type-definition)
with syntax `!seq.clock`, and use it throughout operations. This does
not exist in the CIRCT codebase and clocks are implemented with `i1`.
2. A number of things in `seq` seem to depend on the `hw` dialect which
is not in xdsl at this time. Most CIRCT tests of `seq` things use a
`hw.module({})` region, and `hw.constant` values etc. More concretely,
the attribute `::circt::hw::InnerSymAttr` is used in a number of `seq`
operations. There are a number of ways forward:
a. Define a temporary `InnerSymAttr` in `seq` to be replaced by the
proper import if/when `hw` is implemented,
b. Implement operations using an `InnerSymAttr` without that attribute,
   c. Skip all operations using an `InnerSymAttr`,
   d. Create an empty `hw` dialect with just the needed attribute,
   e. Implement `hw` dialect,
   f. Something I didn’t think of?

---------

Co-authored-by: Chris Vasiladiotis <cvassiladiotis@gmail.com>