interpreter: (riscv) add tracking of register contents (#1631)
This lets us do a few interesting things. One is that we can have a
runtime check of validity of register allocation, as in detect when we
have reused a register in a way that conflicts with the SSA semantics.
Another is that we can now get the contents of a register. Another is
that we can now get the value stored in any register, not just ZERO.