core: Add support for lazy traits in IRDL (#1529)
This is an alternative implementation to #1505, which implements the
laziness at the traits field level, instead of the trait definition
level.
Note that we could force all traits definitions to use `traits_def`,
while right now operations can both use `traits = frozenset(...)` and
`traits = traits_def(frozenset(...))` and `traits = traits_def(lambda:
frozenset(...))`.

I believe this implementation reduces the complexity of the traits
definitions, as they don't need to specify if they are lazy or not. If a
circular dependency arise, this only require a change in the operation
definition, and not the trait definition, which I think is better.