core: (traits) Add `insert_or_update` method to SymbolTable (#1652)
This patch adds a `insert_or_update` method to the SymbolTable trait.
This introduces a portable way of adding symbols if they don't already
exist. As @PapyChacal can confirm, the "if not exists" part is quite
underutilized sometimes.

This patch also refactors two lowerings (printf to putchar, mpi to
function calls) to use the new mechanic. This also fixes the mpi
lowering, which did not previously respect existing symbols with the
given name.