misc: move dependencies out of requirements-optional (#1744)
After reading up a bit on the recommended uses of `requirements.txt`,
`pyproject.toml`, and `setup.py`, I have come to this understanding:

`requirements.txt` is supposed to be used only for local development.
The understanding is that it is a document that is pinned to the local
venv of the developers of the project. It is supposed to be more strict
and specific than the packaging for pypi publishing, which is supposed
to be more permissive.

`pyproject.toml`/`setup.py` are supposed to specify the ranges of
version of dependencies for clients of xdsl that `pip install xdsl` from
their own projects.

It is not recommended to couple the two concepts to each other.

This PR shuffles the same set of dependencies around between files.
`requirements.txt` now holds all the developer dependencies for testing
and linting. People who `pip install xdsl` should not care that their
version of pytest is not the same as ours. `pyproject.toml` now holds
the info of the required dependencies (immutabledict, typing-extensions,
ordered-set), and "extras" dependencies (riscemu, wgpu).

The new command to install xdsl for local development is `pip install -r
requirements.txt` (or `make venv`).

I've tested that `pip install -e .` produces exactly the same output
from `pip list` before and after, and the command above installs exactly
the same packages as `pip install -e ".[extras]"` before. Now the same
command skips the testing projects, which I think is good.

Here are some links for future reference:


https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
https://peps.python.org/pep-0508/
https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html