dialects: (hw) Add InnerSymTarget class to HW dialect (#1758)
InnerSymTarget is a utility class that is used to store the targets of
inner symbols. It is not usable as part of the dialect, but will be used
by classes in `HW`. I’m not sure if there’s much more to test about it
-- it’s really just there to keep a reference to an operation with some
additional metadata.

Standalone PR in an effort to keep them small and reviewable -- see
#1704 for context.