misc: remove flake8 config (#1769)
I accidentally enabled `flake8` on my computer today, to find that we're
in violation of some of the rules. We don't actually specify it as a
dependency, and my understanding is that ruff is now the source of truth
wrt formatting, so we might as well get rid of these lines of spec.