transformations: add interepreter-based contant folding pass (#1599)
Here's a bit of a dopey idea: if we have a way to get the results of
operations in Python, and all the inputs are constants, why not call the
interpreter during the compilation process? This is done in
[Singeli](https://github.com/mlochbaum/Singeli) a part of the BQN
interpreter/compiler specialized for high-performance tensor
computations. They recently discussed this approach on an episode of the
Array Cast and I thought I'd try it out, turns out it took very little
work. Would love to hear your feedback.