core: add ability to parse MLIR IR that drops dialect names from operations (#1840)
Only parsing, no printing. This unlocks us being able to parse MLIR IR
directly output from the `mlir-opt` command, without generic syntax or
"local scope". I needed this to unlock our experiments for the linalg to
frep project, since we use MLIR as the input, and our version of MLIR
had different names for some properties, which messed up the generic
syntax. I think that the main xdsl project would benefit from these
changes also.

In my understanding, this is more permissive than MLIR itself, that has
specific rules for which ops can drop names and which ones cannot. I'm
not sure whether we want to be as strict as them on this.