backend: (riscv) fix function lowering with mixed floats and ints (#1837)
I was under the impression that, in the presence of mixed x and f
registers, the ABI was to have them be numbered according to their index
in the arg list, so for (int, float, int) -> (float, int) it would be
(a0, fa1, a2) -> (fa0, a1). This is not right, the ABI specifies (a0,
fa0, a1) -> (fa0, a0) for the same example. This PR fixes this.

@xdslproject/risc-v-backend