dialects: (onnx) lint onnx.py (#1966)
Two small nits highligthed by VSCode that I missed in my review:

- FloatAttr and IntegerAttr are generic, and Pylance isn't happy when
the generic parameter is not specified. AnyFloatAttr and AnyIntegerAttr
is probably what we want if we want to allow any bitwidth in the types.
Pylance is the version of Pyright built into VSCode, not sure why it
disagrees with the one we have in the CI.
- There was dead code after the `assert False`, not sure why Pyright
didn't bring this up in the CI either, but we should test for it also.