dialects: (snitch_stream) replace snitch_stream generic with streaming region (#1986)
Apologies for the large diff, I'm not sure how to reduce it.

This change simplifies the dialect quite a bit, by decoupling the
iteration logic from the streaming logic. Instead of having a "generic"
op at the snitch_stream level, it's expected that the streaming region
contains all the iteration logic directly, and the only semantic support
it provides is a high-level representation of a chunk of code where
certain streams can be read from or written to.