core: Implement BytesAttr (#2032)
Here's a non-intrusive variant to be able to roundtrip more general
string literals as MLIR does.
The idea is to have another bytes literal, and an extra BytesAttr
builtin attribute, to handle cases breaking Python's `str`.

It parses all string literals, escaping those binary escape sequences as
MLIR does; if the result is Unicode-decodable, then it's a string
literal! Otherwise, it's a BytesAttr!