dialects: (riscv) Canonicalize the loading 0 to `li zero, 0` (#2077)
Note stacked PR

We need to canonicalize getting the 0 value, so that the other
transforms can have simple logic for checking for constant values.
Adding this canonicalization pattern by itself messed with the assembly
generated downstream, hence the PR this one points to, to get the
assembly back to normal. I have run the verilator based on this PR, all
the tests pass, most have a lower cycle count. One has a higher cycle
count, which makes no sense to me, as the only change is skipping the
branch.