dialects: (snitch_stream) reverse bounds and strides in snitch_stream (#2139)
This PR brings the documentation and interpreter implementation of
snitch_stream stride patterns in line with the snitch_stream lowering to
snitch.

This is a PR with only a functional change in the interpreter, but with
consequences on upcoming lowerings from memref_stream to snitch_stream.
When I first added snitch_stream and the corresponding interpreter
functions, I was under the impression that the snrt functions took
bounds and strides from the outermost loop inwards, but it's the
reverse. The distinction is now somewhat arbitrary, since we are
planning to link to snrt any more, and could actually implement the
lowering from snitch_stream to snitch with the inverse order of bounds.
I think that we should stick to the current order of bounds, implement
the complete lowering first, and then see whether switching the order
around would be an improvement.